US15 and US22 Routing done
